Adds validation to footer links
The HTML and CSS itself can be organized and cleaned up further at a later time. All that remains are any fixes that are noticed through normal use of the site. Not all resolutions may be displaying it properly, but based upon teacher's review it's considered stable. Validation will be added in an upcoming commit.